requirements.txt
denonavr==0.7.7
paho-mqtt==1.4.0
